fix(build): wrap search page in suspense
